Rename search-related files
