+++ app/support/DbAdapter/search.js
